TSDK-761 Add add-secret and get-preimage.
add-secret

This feature allows the user to add a secret to the wallet to be
used in a digest lock. The secret can be encoded either in blake2b
or sha256. Sha256 is a bitcoin compatible hash function.This
feature allows the user to add a secret to the wallet to be used
in a digest lock. The secret can be encoded either in blake2b or
sha256. Sha256 is a bitcoin compatible hash function.

get-preimage

The goal of this feature is to allow the user to query the wallet
database for a specific digest text. This is useful to check if
a given transaction can be proved.The goal of this feature is to
allow the user to query the wallet database for a specific digest
text. This is useful to check if a given transaction can be proved.