TSDK-761 Digest support part 2 (#62)
* TSDK-761 Add add-secret and get-preimage.

add-secret

This feature allows the user to add a secret to the wallet to be
used in a digest lock. The secret can be encoded either in blake2b
or sha256. Sha256 is a bitcoin compatible hash function.This
feature allows the user to add a secret to the wallet to be used
in a digest lock. The secret can be encoded either in blake2b or
sha256. Sha256 is a bitcoin compatible hash function.

get-preimage

The goal of this feature is to allow the user to query the wallet
database for a specific digest text. This is useful to check if
a given transaction can be proved.The goal of this feature is to
allow the user to query the wallet database for a specific digest
text. This is useful to check if a given transaction can be proved.

* TSDK-761 Finish implementation

* TSDK-761 Increase timeout for genesis tx

* TSDK-761 Increase timeout for all attempts

* TSDK-761 Increase timeout even more

* TSDK-761 Add more timeout

* TSDK-761 Fix errors and added more timeout

* TSDK-761 Use latest dev version of node

* Update cli/src/it/scala/co/topl/brambl/cli/WalletDigestPropositionTest.scala

Co-authored-by: Diadem <d.shoukralla@topl.me>

* Update cli/src/main/scala/co/topl/brambl/cli/controllers/WalletController.scala

Co-authored-by: Diadem <d.shoukralla@topl.me>

* Update cli/src/main/scala/co/topl/brambl/cli/controllers/WalletController.scala

Co-authored-by: Diadem <d.shoukralla@topl.me>

* Update cli/src/main/scala/co/topl/brambl/cli/controllers/WalletController.scala

Co-authored-by: Diadem <d.shoukralla@topl.me>

* TSDK-761 Add check for length of secret

* TSDK-761 Fix tests

* TSDK-761 Fix integration tests

* TSDK-761 Fix integration tests

* TSDK-761 Fix integration tests

* TSDK-761 Increase version of Bifrost

* Revert "TSDK-761 Increase version of Bifrost"

This reverts commit d78cae1ea26178c78d8ced51a0f31d66231f6787.

---------

Co-authored-by: Diadem <d.shoukralla@topl.me>